Move core types to their own types module
