+++ src/core/initialize.rs
@@ -4,6 +4,7 @@ use core::copy::*;
 use core::copy;
 use core::proc_maps::*;
 use core::ruby_version;
+use core::types::{StackTrace, Process};
 
 use failure::Error;
 use failure::ResultExt;
@@ -11,8 +12,6 @@ use failure::Fail;
 use read_process_memory::*;
 use libc::{c_char, pid_t};
 
-use std::cmp::Ordering;
-use std::fmt;
 use std::time::Duration;
 use std;
 
@@ -44,62 +43,12 @@ pub fn initialize(pid: pid_t) -> Result<StackTraceGetter<ProcessHandle>, Error> 
     })
 }
 
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize, Deserialize)]
-pub struct StackFrame {
-    pub name: String,
-    pub relative_path: String,
-    pub absolute_path: Option<String>,
-    pub lineno: u32,
-}
-
-impl StackFrame {
-    pub fn path(&self) -> &str {
-        match self.absolute_path {
-            Some(ref p) => p.as_ref(),
-            None => self.relative_path.as_ref(),
-        }
-    }
-}
-
-impl Ord for StackFrame {
-    fn cmp(&self, other: &StackFrame) -> Ordering {
-        self.path()
-            .cmp(other.path())
-            .then(self.name.cmp(&other.name))
-            .then(self.lineno.cmp(&other.lineno))
-    }
-}
-
-impl PartialOrd for StackFrame {
-    fn partial_cmp(&self, other: &StackFrame) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize, Deserialize)]
-pub struct StackTrace {
-    pub trace: Vec<StackFrame>,
-    pub pid: Option<pid_t>,
-}
-
-impl StackTrace {
-    pub fn iter(&self) ->  std::slice::Iter<StackFrame> {
-        self.trace.iter()
-    }
-}
-
 // Use a StackTraceGetter to get stack traces
 pub struct StackTraceGetter<T> where T: CopyAddress {
     process: Process<T>,
     current_thread_addr_location: usize,
     stack_trace_function:
         Box<Fn(usize, &Process<T>) -> Result<StackTrace, MemoryCopyError>>,
-}
-
-impl fmt::Display for StackFrame {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{} - {} line {}", self.name, self.path(), self.lineno)
-    }
 }
 
 impl<T> StackTraceGetter<T> where T: CopyAddress {
@@ -110,11 +59,6 @@ impl<T> StackTraceGetter<T> where T: CopyAddress {
             &self.process,
         )
     }
-}
-
-pub struct Process<T> where T: CopyAddress {
-    pub pid: Option<pid_t>,
-    pub source: T,
 }
 
 // Everything below here is private
