+++ src/core/mod.rs
@@ -1,3 +1,4 @@
+pub mod types;
 pub mod initialize;
 mod address_finder;
 pub mod copy;
